rename modal to page
