+++ packages/frinx-workflow-ui/src/pages/workflow-list/WorkflowDefs/DiagramModal/DiagramModal.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import WorkflowDia from '../../executed-workflow-list/details-modal/WorkflowDia/WorkflowDia';
+import WorkflowDia from '../../executed-workflow-list/executed-workflow-detail/WorkflowDia/WorkflowDia';
 import {
   Button,
   Modal,
