+++ packages/frinx-workflow-ui/src/pages/workflow-list/executed-workflow-list/executed-workflow-detail/WorkflowDia/WorkflowDia.js
