+++ packages/frinx-workflow-ui/src/pages/workflow-list/executed-workflow-list/executed-workflow-detail/executed-workflow-detail-tabs/input-output-tab.tsx
