+++ packages/frinx-workflow-ui/src/pages/workflow-list/executed-workflow-list/executed-workflow-detail/executed-workflow-detail-tabs/workflow-json-tab.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Box, Stack, IconButton, Button, Text, Textarea } from '@chakra-ui/react';
 import { CopyIcon } from '@chakra-ui/icons';
-import { ExecutedWorkflowDetailResult, Status } from '../details-modal';
+import { ExecutedWorkflowDetailResult, Status } from '../executed-workflow-detail';
 
 type Props = {
   isEscaped: boolean;
